All files pertaining to state-space analysis
This project outlines the process of implementing the control system with regard to the two-wheeled coupled with a mass system. The document outlines the process, as well as analysis of equations of motion, linearization of the DEs, state space formulation, open and closed loop stability analysis, observability and controllability analysis, state feedback controller, observer design, and observer-based control. These topics serve are fundamental concepts within modern control systems.